silence warning in bands.cpp

+++ src/bands.cpp
@@ -416,7 +416,7 @@ int do_harminv(complex<double> *data, int n, int sampling_rate, double a,
   return 0;
 #else
   double dt = sampling_rate * c / a;
-  int numfreqs = fabs(fmax-fmin)*dt*n*spectral_density; // c.f. 'man harminv'
+  int numfreqs = int(fabs(fmax-fmin)*dt*n*spectral_density); // c.f. harminv
   if (numfreqs > 150) numfreqs = 150; // prevent matrices from getting too big
   if (numfreqs < 2) numfreqs = 2;
